+++ test/files_test.rb
@@ -29,4 +29,19 @@ class FilesTest < Minitest::Test
 
     assert_requested(:delete, 'http://localhost/api/files/yolo')
   end
+
+  def test_tmp_upload
+    file_list = ["yolo/#{__FILE__}"]
+    stub_request(:post, %r{http://localhost/api/files/Aptly__Files-(.+)})
+      .with(headers: {'Content-Type'=>/multipart\/form-data; boundary=-----------RubyMultipartPost.*/})
+      .to_return(body: JSON.generate(file_list))
+    stub_request(:delete, %r{http://localhost/api/files/Aptly__Files-(.+)})
+
+    yielded = false
+    ::Aptly::Files.tmp_upload([__FILE__]) do |dir|
+      assert_includes(dir, 'Aptly__Files-')
+      yielded = true
+    end
+    assert(yielded, 'expected tmp_upload to yield')
+  end
 end