+++ test/repository_test.rb
@@ -77,10 +77,10 @@ class RepositoryTest < Minitest::Test
 
     # NB: the dir name must not include dots or the server won't be able to
     #   handle it
-    stub_request(:post, %r{http://localhost/api/files/Aptly__Repository-([^\.]+)})
+    stub_request(:post, %r{http://localhost/api/files/Aptly__Files-([^\.]+)})
       .with(headers: {'Content-Type'=>/multipart\/form-data; boundary=-----------RubyMultipartPost.*/})
       .to_return(body: '["Aptly__Repository/kitteh.deb"]')
-    stub_request(:post, %r{http://localhost/api/repos/kitten/file/Aptly__Repository-([^\.]*)})
+    stub_request(:post, %r{http://localhost/api/repos/kitten/file/Aptly__Files-([^\.]*)})
       .to_return(body: "{\"FailedFiles\":[],\"Report\":{\"Warnings\":[],\"Added\":[\"gpgmepp_15.08.2+git20151212.1109+15.04-0_source added\"],\"Removed\":[]}}\n")
     stub_request(:delete, %r{http://localhost/api/files/.+})
       .to_return(body: "{}\n")
@@ -127,7 +127,7 @@ class RepositoryTest < Minitest::Test
 
     # Stub some calls we don't care about, these are implicitly called as part
     # of the high level upload methods.
-    stub_request(:post, Regexp.new("#{host_uri}/api/repos/kitten/file/Aptly__Repository-(.*)"))
+    stub_request(:post, Regexp.new("#{host_uri}/api/repos/kitten/file/Aptly__Files-(.*)"))
       .to_return(body: "{\"FailedFiles\":[],\"Report\":{\"Warnings\":[],\"Added\":[\"gpgmepp_15.08.2+git20151212.1109+15.04-0_source added\"],\"Removed\":[]}}\n")
     stub_request(:delete, Regexp.new("#{host_uri}/api/files/.+"))
       .to_return(body: "{}\n")
@@ -166,11 +166,11 @@ class RepositoryTest < Minitest::Test
 
   def test_erroring_upload
     debfile = File.join(__dir__, 'data', 'kitteh.deb')
-    stub_request(:post, %r{http://localhost/api/files/Aptly__Repository-(.*)})
+    stub_request(:post, %r{http://localhost/api/files/Aptly__Files-(.*)})
       .with(headers: {'Content-Type'=>/multipart\/form-data; boundary=-----------RubyMultipartPost.*/})
       .to_return(body: '["Aptly__Repository/kitteh.deb"]')
-    stub_request(:post, %r{http://localhost/api/repos/kitten/file/Aptly__Repository-(.*)})
-      .to_return(body: "{\"FailedFiles\":[\"/home/nci/aptly/upload/brum/kitteh.deb\"],\"Report\":{\"Warnings\":[\"Unable to process /home/nci/aptly/upload/Aptly__Repository-smith-20151217-14879-7cq2c8/gpgmepp_15.08.2+git20151212.1109+15.04.orig.tar.xz: stat /home/nci/aptly/upload/Aptly__Repository-smith-20151217-14879-7cq2c8/gpgmepp_15.08.2+git20151212.1109+15.04.orig.tar.xz: no such file or directory\"],\"Added\":[],\"Removed\":[]}}\n")
+    stub_request(:post, %r{http://localhost/api/repos/kitten/file/Aptly__Files-(.*)})
+      .to_return(body: "{\"FailedFiles\":[\"/home/nci/aptly/upload/brum/kitteh.deb\"],\"Report\":{\"Warnings\":[\"Unable to process /home/nci/aptly/upload/Aptly__Files-smith-20151217-14879-7cq2c8/gpgmepp_15.08.2+git20151212.1109+15.04.orig.tar.xz: stat /home/nci/aptly/upload/Aptly__Files-smith-20151217-14879-7cq2c8/gpgmepp_15.08.2+git20151212.1109+15.04.orig.tar.xz: no such file or directory\"],\"Added\":[],\"Removed\":[]}}\n")
     stub_request(:delete, %r{http://localhost/api/files/.+})
       .to_return(body: "{}\n")
     repo = ::Aptly::Repository.new(::Aptly::Connection.new, Name: 'kitten')